added data cleaner
